+++ capabilities/rbac/targets/nextjs/files/src/app/api/auth/authorization/route.ts
@@ -1,4 +1,4 @@
-import { handleGetAuthorization } from "../../../../core/auth/handlers/get-authorization-handler.js";
+import { handleGetAuthorization } from "../../../../core/authorization/get-authorization-handler.js";
 import { buildAuthHandlerDeps } from "../../../../core/auth/server/route-deps.js";
 
 export const dynamic = "force-dynamic";