+++ capabilities/rbac/targets/nextjs/files/src/features/authorization/use-authorization.ts
@@ -3,8 +3,8 @@
 import { useQuery } from "@tanstack/react-query";
 
 import { toPublicAuthError, type PublicAuthError } from "../../core/auth/session-state.js";
-import { authorizationQueryOptions } from "./auth-queries.js";
-import { useSession } from "./use-session.js";
+import { authorizationQueryOptions } from "../../core/authorization/authorization-queries.js";
+import { useSession } from "../auth/use-session.js";
 
 export interface UseAuthorizationResult {
   readonly roles: readonly string[];